+++ CHANGELOG.md
@@ -4,6 +4,34 @@ All notable updates to the project will be documented in this file.
 
 ## [Unreleased]
 
+## [Version 1.1.0] - 2023-09-16
+
+### New Features
+
+New ErrorAnalysisEnhancer class for error stack analysis and extraction of file, line, column info etc.
+New SerializersUtility class to serialize errors to JSON, CSV, XML and YAML
+Added pattern to capture originating module and method context
+
+### Enhancements
+
+Refactored validators into standalone files
+HttpStatusEnhancer now takes number instead of enum for status code
+IdentifiersEnhancer now takes strings instead of enums for severity and category
+Updated examples
+
+### Bug Fixes
+
+Fixed issue capturing system context in SystemContextEnhancer
+
+### Other changes
+
+Removed individual serializer classes in favor of SerializersUtility
+Removed FilterHelper in favor of FilterUtility
+Removed unused validators
+Improved package.json metadata
+Switched to ESM, added exports field
+Added JSDoc comments
+
 ## [Version 1.0.0] - 2023-09-15
 
 ### New Features
@@ -32,7 +60,7 @@ All notable updates to the project will be documented in this file.
 
 ### Authors
 
-- Alessandro Barbagallo
+- alessbarb
 
 [Unreleased]: https://github.com/labrynx/error-enhanced/compare/previous-version...HEAD
 [Version 1.0.0]: https://github.com/labrynx/error-enhanced/releases/tag/version-1.0.0